Convert AppManagementFilter to TS components
